Native: add file extension and option to kill process when leave the app

+++ source/Native Application/DMCNativeGUI_V4.0/src/com/app/dlna/dmc/processor/upnp/CoreUpnpService.java
@@ -47,6 +47,7 @@ import android.util.Log;
 import android.widget.Toast;
 import app.dlna.controller.v4.R;
 
+import com.app.dlna.dmc.gui.activity.AppPreference;
 import com.app.dlna.dmc.gui.activity.MainActivity;
 import com.app.dlna.dmc.processor.cache.Cache;
 import com.app.dlna.dmc.processor.http.HTTPServerData;
@@ -234,7 +235,8 @@ public class CoreUpnpService extends Service {
 				protected void onPostExecute(Void result) {
 					if (m_notificationManager != null)
 						m_notificationManager.cancel(NOTIFICATION);
-					android.os.Process.killProcess(android.os.Process.myPid());
+					if (AppPreference.getKillProcessStatus())
+						android.os.Process.killProcess(android.os.Process.myPid());
 				};
 			}.execute(new Void[] {});
 		}
